Rename references (#47)
* Update refs from python_wsdb to mapepire_python


* Update mapepire_python imports to use relative paths


* Update remaining python_wsdb refs to mapepire_python


* Rename refs from python-wsdb to mapepire-python


* Formatting and linting


* Fix import error


* Update CHANGELOG


---------
