+++ .github/pull_request_template.md
@@ -11,9 +11,9 @@ Changes proposed in this pull request:
 
 <!-- Please complete this checklist BEFORE submitting your PR to speed along the review process. -->
 - [ ] Change the base branch to `dev` if it is not already.
-- [ ] I've read and followed all steps in the [Making a pull request](https://github.com/ajshedivy/python-wsdb/blob/main/.github/CONTRIBUTING.md#making-a-pull-request)
+- [ ] I've read and followed all steps in the [Making a pull request](https://github.com/Mapepire-IBMi/mapepire-python/blob/main/.github/CONTRIBUTING.md#making-a-pull-request)
     section of the `CONTRIBUTING` docs.
 - [ ] I've updated or added any relevant docstrings following the syntax described in the
-    [Writing docstrings](https://github.com/ajshedivy/python-wsdb/blob/main/.github/CONTRIBUTING.md#writing-docstrings) section of the `CONTRIBUTING` docs.
+    [Writing docstrings](https://github.com/Mapepire-IBMi/mapepire-python/blob/main/.github/CONTRIBUTING.md#writing-docstrings) section of the `CONTRIBUTING` docs.
 - [ ] If this PR fixes a bug, I've added a test that will fail without my fix.
 - [ ] If this PR adds a new feature, I've added tests that sufficiently cover my new functionality.
