+++ CHANGELOG.md
@@ -7,15 +7,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## [v0.1.4] - 2025-08-09
+Rename references from python_wsdb and python-wsdb to mapepire_python and mapepire-python
+
 ## [v0.1.3] - 2025-08-02
 Add workflow to run test suite on PR's
 Update requirements-dev.txt to include isort, black, ruff, mypy
 
-## [v0.1.2](https://github.com/ajshedivy/python-wsdb/releases/tag/v0.1.2) - 2024-04-22
+## [v0.1.2](https://github.com/Mapepire-IBMi/mapepire-python/releases/tag/v0.1.2) - 2024-04-22
 
-## [v0.1.1](https://github.com/ajshedivy/python-wsdb/releases/tag/v0.1.1) - 2024-04-19
+## [v0.1.1](https://github.com/Mapepire-IBMi/mapepire-python/releases/tag/v0.1.1) - 2024-04-19
 
-## [v0.1.0](https://github.com/ajshedivy/python-wsdb/releases/tag/v0.1.0) - 2024-04-19
+## [v0.1.0](https://github.com/Mapepire-IBMi/mapepire-python/releases/tag/v0.1.0) - 2024-04-19
 Add initial release
 
 ## [0.1.2]
