+++ docs/source/installation.md
@@ -1,11 +1,11 @@
 Installation
 ============
 
-**python-wsdb** supports Python >= 3.8.
+**mapepire-python** supports Python >= 3.8.
 
 ## Installing with `pip`
 
-**python-wsdb** is available [on PyPI](https://pypi.org/project/python-wsdb/). Just run
+**mapepire-python** is available [on PyPI](https://pypi.org/project/python-wsdb/). Just run
 
 ```bash
 pip install python-wsdb
@@ -13,11 +13,11 @@ pip install python-wsdb
 
 ## Installing from source
 
-To install **python-wsdb** from source, first clone [the repository](https://github.com/ajshedivy/python-wsdb):
+To install **mapepire-python** from source, first clone [the repository](https://github.com/Mapepire-IBMi/mapepire-python):
 
 ```bash
-git clone https://github.com/ajshedivy/python-wsdb.git
-cd python-wsdb
+git clone https://github.com/Mapepire-IBMi/mapepire-python
+cd mapepire-python
 ```
 
 Then run
