+++ mapepire_python/client/query.py
@@ -2,8 +2,8 @@ import json
 from enum import Enum
 from typing import Any, Dict, Generic, List, Optional, TypeVar, Union
 
-from python_wsdb.client.sql_job import SQLJob
-from python_wsdb.types import QueryOptions
+from ..types import QueryOptions
+from .sql_job import SQLJob
 
 T = TypeVar("T")
