+++ mapepire_python/ssl.py
@@ -2,7 +2,7 @@ import socket
 import ssl
 from typing import Optional
 
-from python_wsdb.types import DaemonServer
+from .types import DaemonServer
 
 
 def get_certificate(creds: DaemonServer) -> Optional[bytes]:
