--- python_wsdb/__init__.py
@@ -1,1 +1,0 @@
-from .version import VERSION, VERSION_SHORT
